refactor

+++ users/signals.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 from .models import Profile
 
 
-def createProfileAfterCreateUser(sender, instance, created, **kwargs):
+def create_profile_after_create_user(sender, instance, created, **kwargs):
     if created:
         user = instance
         profile = Profile.objects.create(
@@ -13,12 +13,26 @@ def createProfileAfterCreateUser(sender, instance, created, **kwargs):
             email=user.email,
             name=user.first_name,
         )
-        print('Profile created!')
-
-
-def deleteUserAfterDeleteProfile(sender, instance, **kwargs):
-    user = instance.user
-    user.delete()
+        
+        
+def update_user_after_update_profile(sender, instance, created, **kwargs):
+    if not instance.user:
+        return
     
-post_save.connect(createProfileAfterCreateUser, sender=User)
-post_delete.connect(deleteUserAfterDeleteProfile, sender=Profile)
+    user = instance.user
+    if not created:
+        user.username = instance.username
+        user.email = instance.email
+        user.first_name = instance.name
+        user.save()
+
+
+def delete_user_after_delete_profile(sender, instance, **kwargs):
+    if instance.user:
+        user = instance.user
+        user.delete()
+    
+    
+post_save.connect(update_user_after_update_profile, sender=Profile)
+post_save.connect(create_profile_after_create_user, sender=User)
+post_delete.connect(delete_user_after_delete_profile, sender=Profile)
